Broke out filter and form for re-use
